fix(box2d): extend the no-FMA guarantee to test code
The previous commit removed fusion from the package but the amd64
golden tests still failed. Golden values are partly computed by test
code -- scene layout, seeded input tables, the spring integration loop
-- and `go build` never compiles test files, so the gate could not see
those sites. Three goldens diverged for exactly that reason.

Round the products in those computations, including one that is only a
product after the compiler strength-reduces a division by a power of
two into a multiply. Compile the test binary in the gate as well, so a
fusion introduced by a future test is caught the same way.

arm64 now emits no FMA instruction in either the package or the test
binary, and goldens verify clean on repeated runs.

+++ pkg/box2d/fuzz_ops_test.go
@@ -141,7 +141,7 @@ func srcUnit(s opSource) float64 {
 
 // srcRange returns a value in [lo, hi).
 func srcRange(s opSource, lo, hi float64) float64 {
-	return lo + (hi-lo)*srcUnit(s)
+	return lo + float64((hi-lo)*srcUnit(s))
 }
 
 // ---------------------------------------------------------------------------
